CHANGED SELECTORS AND REPLACED THE LAST ANIMATION FOR ANOTHER
Line 57 to 58 were changed the selectors bc in the HTML were changed the classes for a better seletion time.

Line 66 to 69 was changed the position to relative for <a> labels in the navbar, this was done to avoid changing the style in the background.

Line 71 to 73 was replaced the last animation for a border_animation.

from line 91 to 104 was changed the instructions for the animation and now it creates a border. IDEA: Start from a point or a short border and then expands at the 100% width of the <a> label.